Refactored imports and utilities by splitting `utils.ts` into separate modules for better organization and maintainability.
- Removed `src/utils.ts` and distributed its functionalities into:
  - `src/utils/general.ts` for general utility functions like `getBoundary`, `getCharsetName`, and `mimeDecode`.
  - `src/utils/guid.ts` for `guid` generation.
  - `src/utils/gbkutf.ts` for encoding functions related to GB2312 to UTF-8 conversions.
  - `src/utils/string.ts` for string manipulations like `wrap` and `isStringOrError`.

- Updated corresponding import paths in `src/index.ts` to reflect the new structure.
- Modified `read()` function in `src/index.ts` to use the new `GB2312UTF8` function directly.
- Deleted `test/util.test.ts` and created new test files aligned with the new module structure:
  - `test/utils/general.test.ts`
  - `test/utils/guid.test.ts`
  - `test/utils/gbkutf.test.ts`
  - `test/utils/string.test.ts`

These changes enhance code modularity and make utility functions more accessible and easier to maintain.
