+++ src/utils/guid.ts
@@ -1,0 +1,10 @@
+//Generates a random id
+export function guid() {
+    return 'xxxxxxxxxxxx-4xxx-yxxx-xxxxxxxxxxxx'
+        .replace(/[xy]/g, function (c) {
+            const r = (Math.random() * 16) | 0,
+                v = c == 'x' ? r : (r & 0x3) | 0x8;
+            return v.toString(16);
+        })
+        .replace('-', '');
+}
